stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
# Установка gunicorn
gunicorn
django-extensions
django_filter